third_party/WebKit/LayoutTests/fast/events/gesture-pinch-zoom-expected.txt

PASS window.innerWidth is 800
PASS window.innerHeight is 600
PASS gotWheelEvent is false
PASS event.ctrlKey is true
PASS event.deltaX is within 0.001 of 0
PASS event.deltaY is within 0.001 of -40.54651081081644
PASS event.x is within 0.001 of 50
PASS event.y is within 0.001 of 40
PASS window.innerWidth is 800
PASS window.innerHeight is 600
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS gotWheelEvent is true
PASS gotWheelEvent is false
PASS internals.visualViewportWidth() is 400
PASS internals.visualViewportHeight() is 300
PASS internals.visualViewportScrollX() is 75
PASS internals.visualViewportScrollY() is 50
PASS successfullyParsed is true

TEST COMPLETE
